requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1
ReportLab==4.0.7
python-docx==1.1.0
python-dotenv==1.0.0
gunicorn
